Added size output

+++ team.h
@@ -10,7 +10,7 @@ public:
 	team();
     float recalc_team_avg_fit();
 	float calc_team_fit();
-    float write_team_fit_final(std::ofstream&, int, int, int, float);
+	float write_team_fit_final(std::ofstream&, int, int, int, int);
 	float get_team_fit() {
 		return this->team_fit;
 	}
